+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changes
 
 ## 0.2.1
+* Explicitly call del and gc during update to better free memory
 * InstanceV tracks nodes_popped, InstanceQ tracks edges_popped, more memory efficient updater
 * map_location=torch.device('cpu') when loading on cpu
 * DataParallel during solve